// src/app/features/teams/enums/team-name.ts
export enum TeamName {
  GentleMates = 'Paris Gentle Mates',
  Faze = 'Faze Vegas',
  Falcons = 'Riyadh Falcons',
  KOI = 'Toronto KOI',
  Surge = 'Vancouver Surge',
  Optic = 'Optic Texas',
  Heretics = 'Miami Heretics',
  Thieves = 'Los Angeles Thieves',
  G2 = 'G2 Minnesota',
  Cloud9 = 'Cloud9 New York',
  RoyalRavens = 'Carolina Royal Ravens',
  BostonBreach = 'Boston Breach',
}